prng: add function to calculate pareto value from percentile

+++ src/test.cpp
@@ -1337,7 +1337,7 @@ void test_bounded_pareto()
     constexpr auto kNumbersToGenerate = 10'000'000;
     double min = INFINITY;
     double max = 0;
-    long double mean = 0;
+    double mean = 0;
 
     for (auto i = 0; i < kNumbersToGenerate; ++i) {
         auto generated_number = generator.generate();
